Move JSON schema based validator to mds-schema-validators
