+++ .vscode/launch.json
@@ -912,6 +912,26 @@
       }
     },
     {
+      "name": "Tests: mds-schema-validators",
+      "type": "node",
+      "request": "launch",
+      "runtimeArgs": [
+        "--inspect-brk",
+        "${workspaceRoot}/node_modules/jest/bin/jest",
+        "--runInBand",
+        "--coverage",
+        "false"
+      ],
+      "console": "integratedTerminal",
+      "internalConsoleOptions": "neverOpen",
+      "protocol": "inspector",
+      "cwd": "${workspaceFolder}/packages/mds-schema-validators",
+      "env": {
+        "DOTENV_CONFIG_PATH": "${workspaceFolder}/.env",
+        "npm_package_name": "@mds-core/mds-schema-validators"
+      }
+    },
+    {
       "name": "Tests: mds-stream",
       "type": "node",
       "request": "launch",
