+++ packages/mds-collector-api/package.json
@@ -24,6 +24,7 @@
     "@mds-core/mds-api-server": "0.1.28",
     "@mds-core/mds-collector-backend": "0.1.2",
     "@mds-core/mds-logger": "0.2.2",
+    "@mds-core/mds-schema-validators": "0.1.4",
     "@mds-core/mds-service-helpers": "0.1.2",
     "@mds-core/mds-types": "0.1.25",
     "@mds-core/mds-utils": "0.1.28",
