+++ packages/mds-collector-api/tests/api.spec.ts
@@ -18,12 +18,12 @@ import supertest from 'supertest'
 import HttpStatus from 'http-status-codes'
 import { CollectorServiceClient } from '@mds-core/mds-collector-backend'
 import { CollectorBackendController } from '@mds-core/mds-collector-backend/service/backend'
+import { TestSchema } from '@mds-core/mds-schema-validators'
 import { ApiServer } from '@mds-core/mds-api-server'
 import { pathPrefix, uuid } from '@mds-core/mds-utils'
 import { UUID } from '@mds-core/mds-types'
 import { CollectorApiAccessTokenScopes, COLLECTOR_API_DEFAULT_VERSION, COLLECTOR_API_MIME_TYPE } from '../@types'
 import { api } from '../api'
-import TestSchema from '@mds-core/mds-collector-backend/schemas/test.schema'
 
 const CollectorBackend = CollectorBackendController()
 const request = supertest(ApiServer(api))
