+++ packages/mds-collector-api/tsconfig.build.json
@@ -8,6 +8,7 @@
     { "path": "../../packages/mds-api-server/tsconfig.build.json" },
     { "path": "../../packages/mds-collector-backend/tsconfig.build.json" },
     { "path": "../../packages/mds-logger/tsconfig.build.json" },
+    { "path": "../../packages/mds-schema-validators/tsconfig.build.json" },
     { "path": "../../packages/mds-service-helpers/tsconfig.build.json" },
     { "path": "../../packages/mds-types/tsconfig.build.json" },
     { "path": "../../packages/mds-utils/tsconfig.build.json" }
