+++ packages/mds-collector-backend/index.ts
@@ -16,4 +16,3 @@
 
 export * from './@types'
 export * from './client'
-export * from './schema-validator'
