+++ packages/mds-collector-backend/package.json
@@ -25,12 +25,12 @@
     "@mds-core/mds-logger": "0.2.2",
     "@mds-core/mds-repository": "0.1.2",
     "@mds-core/mds-rpc-common": "0.1.2",
+    "@mds-core/mds-schema-validators": "0.1.4",
     "@mds-core/mds-service-helpers": "0.1.2",
     "@mds-core/mds-stream": "0.1.28",
     "@mds-core/mds-types": "0.1.25",
     "@mds-core/mds-utils": "0.1.28",
     "ajv": "8.0.0",
-    "ajv-formats": "2.0.0",
     "typeorm": "0.2.30"
   }
 }
