+++ packages/mds-schema-validators/jest.config.js
@@ -1,0 +1,11 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const baseConfig = require('../../jest.config')
+
+module.exports = {
+  ...baseConfig,
+  coverageThreshold: {
+    global: {
+      lines: 75
+    }
+  }
+}
