+++ packages/mds-schema-validators/package.json
@@ -16,12 +16,14 @@
   ],
   "scripts": {
     "build": "tsc --build tsconfig.build.json",
-    "test": "DOTENV_CONFIG_PATH=../../.env nyc --use-spawn-wrap --lines 75  ts-mocha --project ../../tsconfig.json"
+    "test": "DOTENV_CONFIG_PATH=../../.env jest"
   },
   "dependencies": {
     "@mds-core/mds-providers": "0.1.28",
     "@mds-core/mds-types": "0.1.25",
     "@mds-core/mds-utils": "0.1.28",
+    "ajv": "8.0.3",
+    "ajv-formats": "2.0.1",
     "joi": "17.3.0",
     "joi-to-json": "1.2.0"
   }
